+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Upcoming
 
+- **ADD:** Support namespaces for dependency matching.
+- **CHANGE:** Print full module name instead of top level only.
+
 ## [1.5.4]
 - **FIX:** Set Author and Email for release version tag pipeline